Fix magic mirror (#28084)
* Fix magic mirror

* buff magic mirror
