+++ Content.Shared/MagicMirror/MagicMirrorComponent.cs
@@ -47,5 +47,5 @@ public sealed partial class MagicMirrorComponent : Component
     /// Sound emitted when slots are changed
     /// </summary>
     [DataField, ViewVariables(VVAccess.ReadWrite)]
-    public SoundSpecifier ChangeHairSound = new SoundPathSpecifier("/Audio/Items/scissors.ogg");
+    public SoundSpecifier? ChangeHairSound = new SoundPathSpecifier("/Audio/Items/scissors.ogg");
 }
