+++ Resources/Prototypes/Entities/Structures/Wallmounts/mirror.yml
@@ -11,7 +11,12 @@
   - type: Clickable
   - type: Transform
     anchored: true
-  - type: MagicMirror
+  - type: MagicMirror #instant and silent
+    changeHairSound: null
+    addSlotTime: 0
+    removeSlotTime: 0
+    selectSlotTime: 0
+    changeSlotTime: 0
   - type: ActivatableUI
     key: enum.MagicMirrorUiKey.Key
   - type: UserInterface
